# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projects Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Projects API Server"
        }
    ],
    "paths": {
...
        "/tasks": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Store a new task",
                "description": "Returns created task data",
                "operationId": "storeTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "project_id": {
                                        "$ref": "#components/schemas/Project/properties/id"
                                    },
                                    "title": {
                                        "$ref": "#components/schemas/Task/properties/title"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "title": "data",
                                            "description": "Data-Wrapper",
                                            "properties": {
                                                "id": {
                                                    "$ref": "#components/schemas/Task/properties/id"
                                                },
                                                "title": {
                                                    "$ref": "#components/schemas/Task/properties/title"
                                                },
                                                "project_id": {
                                                    "$ref": "#components/schemas/Task/properties/project_id"
                                                },
                                                "created_at": {
                                                    "$ref": "#components/schemas/Task/properties/created_at"
                                                },
                                                "updated_at": {
                                                    "$ref": "#components/schemas/Task/properties/updated_at"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{taskId}": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Mark a task as complete",
                "description": "Returns updated task data with is_completed to true",
                "operationId": "markTaskAsComplete",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "title": "data",
                                            "description": "Data-Wrapper",
                                            "properties": {
                                                "id": {
                                                    "$ref": "#components/schemas/Task/properties/id"
                                                },
                                                "title": {
                                                    "$ref": "#components/schemas/Task/properties/title"
                                                },
                                                "project_id": {
                                                    "$ref": "#components/schemas/Task/properties/project_id"
                                                },
                                                "is_completed": {
                                                    "$ref": "#components/schemas/Task/properties/is_completed"
                                                },
                                                "created_at": {
                                                    "$ref": "#components/schemas/Task/properties/created_at"
                                                },
                                                "updated_at": {
                                                    "$ref": "#components/schemas/Task/properties/updated_at"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}